# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5dc0d9f09ef7a1181f8e133e
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc0dceb9ef7a1181f8e1346
  title: demo
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  User  -----
  /user/{id}/changePassword:
    post:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1359
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles: ["ADMIN"]
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
    
  /user/:
    post:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1358
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1355
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    
  /user/{id}:
    delete:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1354
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e135a
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1353
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  sanpham  -----
  /sanpham/{id}:
    get:
      x-skaffolder-id: 5dc0ef509ef7a1181f8e1479
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0ef469ef7a1181f8e1477
      x-skaffolder-resource: sanpham
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: sanpham
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - sanpham
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id sanpham
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sanpham"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:

  # Security
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
        

    User:
      x-skaffolder-id: 5dc0dcec9ef7a1181f8e1350
      x-skaffolder-id-db: 5dc0dceb9ef7a1181f8e1347
      x-skaffolder-id-entity: 5dc0dceb9ef7a1181f8e1348
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e134b
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e134c
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e134a
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e134e
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e134d
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dc0dceb9ef7a1181f8e1349
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    sanpham:
      x-skaffolder-id: 5dc0ef469ef7a1181f8e1477
      x-skaffolder-id-db: 5dc0dceb9ef7a1181f8e1347
      x-skaffolder-id-entity: 5dc0ef469ef7a1181f8e1474
      x-skaffolder-url: /sanpham
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        tensanpham:
          type: string
          x-skaffolder-id-attr: 5dc0ef469ef7a1181f8e1475
          x-skaffolder-type: String
      x-skaffolder-relations: 
        fs:
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 5dc0ef469ef7a1181f8e1474
          x-skaffolder-ent2: 5dc0dceb9ef7a1181f8e1348
      required:
        - _id
      
    
    
  # Databases
  x-skaffolder-db:
    - x-skaffolder-id: 5dc0dceb9ef7a1181f8e1347
      x-skaffolder-name: demo_db

  # Pages
  x-skaffolder-page:
    - x-skaffolder-id: 5dc0dceb9ef7a1181f8e134f
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []